fix(review): Move impementation class into its own file, namespace template comments
