+++ modules/new_serverpod_auth/serverpod_auth_email/serverpod_auth_email_server/lib/serverpod_auth_email_server.dart
@@ -7,6 +7,8 @@ export 'package:serverpod_auth_session_server/serverpod_auth_session_server.dart
 export 'package:serverpod_auth_user_server/serverpod_auth_user_server.dart'
     hide Endpoints, Protocol;
 
+export 'src/business/email_account_implementation.dart'
+    show EmailAccountImplementation;
 export 'src/endpoints/email_account_base_endpoint.dart'
     show EmailAccountBaseEndpoint;
 export 'src/generated/endpoints.dart';
